Feature/follow -> BE

See merge request s08-bigdata-recom-sub2/S08P22A505!43

+++ backend/billboard/src/main/java/com/ssafy/billboard/model/entity/Room.java
@@ -17,7 +17,7 @@ public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, updatable = false)
-    private int roomId;
+    private long roomId;
 
     @Column(nullable = false)
     private String hostId;
